Implemented endpoints

+++ app/services/book_service.py
@@ -1,5 +1,5 @@
 from sqlalchemy.orm import Session
-from ..database.models import Book as BookModel
+from ..database.models import Book as BookModel, BookType as BookTypeModel
 
 
 def get_all_books(db: Session) -> object:
@@ -14,15 +14,43 @@ def get_book_by_id(db: Session, book_id: int) -> object:
 
 
 def create_book(db: Session, book: dict):
-    book = BookModel(**book.dict())
-    db.add(book)
+    # Check if the book type exists
+    db_book_type = db.query(BookTypeModel).filter(BookTypeModel.id == book.book_type_id).first()
+    if db_book_type is None:
+        raise ValueError("Book type not found!")
+
+    # Check if the optional id field is exists
+    if book.id is not None:
+        # Check if a book with the same id already exists
+        db_book_id = db.query(BookModel).filter(BookModel.id == book.id).first()
+        if db_book_id is not None:
+            raise ValueError("Appointed id already exists!")
+
+    # Check if a book with the same name and author already exists
+    db_book = db.query(BookModel).filter(BookModel.name == book.name, BookModel.author == book.author).first()
+
+    # If a book with the same name and author exists, raise an error
+    if db_book is not None:
+        raise ValueError("Book already exists!")
+
+    db_book = BookModel(**book.dict())
+    db.add(db_book)
     db.commit()
-    db.refresh(book)
-    return book
+    db.refresh(db_book)
+
+    return db_book
 
 
 def update_book(db: Session, book_id: int, book):
     db_book = db.query(BookModel).filter(BookModel.id == book_id).first()
+
+    if db_book is None:
+        raise ValueError("Book not found!")
+    db_book_type = db.query(BookTypeModel).filter(BookTypeModel.id == book.book_type_id).first()
+
+    if db_book_type is None:
+        raise ValueError("Book type not found!")
+
     for key, value in book.dict().items():
         setattr(db_book, key, value)
     db.commit()
@@ -32,6 +60,10 @@ def update_book(db: Session, book_id: int, book):
 
 def delete_book(db: Session, book_id: int):
     db_book = db.query(BookModel).filter(BookModel.id == book_id).first()
+
+    if db_book is None:
+        raise ValueError("Book not found!")
+
     db.delete(db_book)
     db.commit()
     return db_book
